Move 'IntChange' to the new 'pkg/state/internal' package.
