+++ pkg/state/diff_applier_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/wavesplatform/gowaves/pkg/proto"
 	"github.com/wavesplatform/gowaves/pkg/settings"
-	"github.com/wavesplatform/gowaves/pkg/util/common"
+	"github.com/wavesplatform/gowaves/pkg/state/internal"
 )
 
 type diffApplierTestObjects struct {
@@ -49,14 +49,14 @@ func TestDiffApplierWithWaves(t *testing.T) {
 	err = to.applier.applyBalancesChanges(changes)
 	assert.Error(t, err, "applyBalancesChanges() did not fail with balance change leading to negative balance")
 	// Test applying invalid leasing change.
-	diff = balanceDiff{leaseOut: common.NewIntChange[int64](101), blockID: blockID0}
+	diff = balanceDiff{leaseOut: internal.NewIntChange[int64](101), blockID: blockID0}
 	changes = []balanceChanges{
 		{[]byte(testGlobal.senderInfo.wavesKey), []balanceDiff{diff}},
 	}
 	err = to.applier.applyBalancesChanges(changes)
 	assert.Error(t, err, "applyBalancesChanges() did not fail with leasing change leading to negative balance")
 	// Valid leasing change.
-	diff = balanceDiff{leaseIn: common.NewIntChange[int64](10), blockID: blockID0}
+	diff = balanceDiff{leaseIn: internal.NewIntChange[int64](10), blockID: blockID0}
 	changes = []balanceChanges{
 		{[]byte(testGlobal.senderInfo.wavesKey), []balanceDiff{diff}},
 	}
@@ -67,7 +67,7 @@ func TestDiffApplierWithWaves(t *testing.T) {
 	assert.NoError(t, err, "wavesBalance() failed")
 	assert.Equal(t, diff.leaseIn.Value(), profile.leaseIn)
 	// Test that leasing leased money leads to error.
-	diff = balanceDiff{leaseOut: common.NewIntChange[int64](101), blockID: blockID0}
+	diff = balanceDiff{leaseOut: internal.NewIntChange[int64](101), blockID: blockID0}
 	changes = []balanceChanges{
 		{[]byte(testGlobal.senderInfo.wavesKey), []balanceDiff{diff}},
 	}
