+++ pkg/state/leases.go
@@ -11,7 +11,7 @@ import (
 	"github.com/wavesplatform/gowaves/pkg/crypto"
 	"github.com/wavesplatform/gowaves/pkg/keyvalue"
 	"github.com/wavesplatform/gowaves/pkg/proto"
-	"github.com/wavesplatform/gowaves/pkg/util/common"
+	"github.com/wavesplatform/gowaves/pkg/state/internal"
 )
 
 type LeaseStatus byte
@@ -159,7 +159,7 @@ func (l *leases) cancelLeasesToDisabledAliases(
 			return nil, nil, errors.Wrapf(err, "failed to build address for PK %q", record.SenderPK)
 		}
 		if diff, ok := changes[senderAddr]; ok {
-			newLeaseOut, loErr := diff.leaseOut.Add(common.NewIntChange(-int64(record.Amount)))
+			newLeaseOut, loErr := diff.leaseOut.Add(internal.NewIntChange(-int64(record.Amount)))
 			if loErr != nil {
 				return nil, nil, errors.Wrapf(loErr, "failed to add leaseOut change for address %q",
 					senderAddr.String(),
@@ -171,7 +171,7 @@ func (l *leases) cancelLeasesToDisabledAliases(
 			changes[senderAddr] = newBalanceDiff(0, 0, -int64(record.Amount), false)
 		}
 		if diff, ok := changes[record.RecipientAddr]; ok {
-			newLeaseIn, liErr := diff.leaseIn.Add(common.NewIntChange(-int64(record.Amount)))
+			newLeaseIn, liErr := diff.leaseIn.Add(internal.NewIntChange(-int64(record.Amount)))
 			if liErr != nil {
 				return nil, nil, errors.Wrapf(liErr, "failed to add leaseIn change for address %q",
 					record.RecipientAddr.String(),
